feat(UI-1239): sessions filter - refactor
