+++ src/components/molecules/popover/index.ts
@@ -2,4 +2,5 @@ export { PopoverWrapper } from "@components/molecules/popover/popover";
 export { PopoverContent, PopoverTrigger } from "@components/molecules/popover/popoverContent";
 export { PopoverContentBase } from "@components/molecules/popover/popoverContentBase";
 export { PopoverListWrapper } from "@components/molecules/popover/popoverList";
-export { PopoverListContent, PopoverListTrigger } from "@components/molecules/popover/popoverListContent";
+export { PopoverListContent } from "@components/molecules/popover/popoverListContent";
+export { PopoverListTrigger } from "@components/molecules/popover/popoverListTrigger";
